+++ base/docs/basedocs.jl
@@ -3020,7 +3020,7 @@ QuoteNode
 
 """
     "
-`"` Is used to delimit string literals.
+`"` Is used to delimit string literals. A trailing `\\` can be used to continue a string literal on the next line.
 
 # Examples
 
@@ -3030,6 +3030,10 @@ julia> "Hello World!"
 
 julia> "Hello World!\\n"
 "Hello World!\\n"
+
+julia> "Hello \\
+        World"
+"Hello World"
 ```
 
 See also [`\"""`](@ref \"\"\").