+++ test/staged.jl
@@ -305,3 +305,30 @@ end
 end
 @test f33243() === 2
 @test x33243 === 2
+
+# https://github.com/JuliaDebug/CassetteOverlay.jl/issues/12
+# generated function with varargs and unfortunately placed unused slot
+@generated function f_vararg_generated(args...)
+    :($args)
+end
+g_vararg_generated() = f_vararg_generated((;), (;), Base.inferencebarrier((;)))
+let tup = g_vararg_generated()
+    @test all(==(typeof((;))), tup)
+    # This is just to make sure that the test is actually testing what we want -
+    # the test only works if there's an unused that matches the position of the
+    # inferencebarrier argument above (N.B. the generator function itself
+    # shifts everything over by 1)
+    @test only(code_lowered(only(methods(f_vararg_generated)).generator.gen)).slotflags[5] == UInt8(0x00)
+end
+
+# respect a given linetable in code generation
+# https://github.com/JuliaLang/julia/pull/47750
+let match = Base._which(Tuple{typeof(sin),Int})
+    mi = Core.Compiler.specialize_method(match)
+    lwr = Core.Compiler.retrieve_code_info(mi)
+    @test all(lin->lin.method===:sin, lwr.linetable)
+    @generated sin_generated(a) = lwr
+    src = only(code_lowered(sin_generated, (Int,)))
+    @test all(lin->lin.method===:sin, src.linetable)
+    @test sin_generated(42) == sin(42)
+end